+++ CHANGELOG.md
@@ -8,6 +8,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Version 0.2.2 - Unreleased
 
 
+### Added
+
+* Auxiliary images now have `warp_aux_to_img` property
+* Base images now have `warp_img_to_vid` property
+
+### Changed
+
+* Removed `base_to_aux` transform in favor of inverse `warp_aux_to_img`
+
+
 ## Version 0.2.1 - Released 2021-04-26